Now able to process input of the mouse and Keyboard. So the player can look/move around in the field.

+++ src/main/java/com/github/kajdreef/mazerunnermvn/State/MainMenu.java
@@ -28,7 +28,7 @@ public class MainMenu extends State {
     }
     
     @Override
-    public void logic(int delta) {
+    public void logic(float delta) {
 
     }
